feat(universal-modal): rework sizing
