+++ packages/universal-modal/src/desktop/components/center-modal/center-modal.module.css
@@ -2,7 +2,6 @@
 
 .component {
     min-width: 500px;
-    width: 500px;
     min-height: 264px;
     border-radius: var(--border-radius-24);
     overflow-x: hidden;
